refactor: Move `AssignFailEvent` and `AssignSuccessEvent` to `deliverymanager.hub` package
