+++ delivery-service/src/main/java/com/winner/client/deliveryservice/common/event/deliverymanager/hub/AssignFailEvent.java
@@ -1,4 +1,4 @@
-package com.winner.client.deliveryservice.common.event;
+package com.winner.client.deliveryservice.common.event.deliverymanager.hub;
 
 public record AssignFailEvent(
 	String message
