+++ delivery-service/src/main/java/com/winner/client/deliveryservice/common/event/deliverymanager/hub/AssignSuccessEvent.java
@@ -1,4 +1,4 @@
-package com.winner.client.deliveryservice.common.event;
+package com.winner.client.deliveryservice.common.event.deliverymanager.hub;
 
 import java.util.UUID;
